+++ lib/combo/template.ex
@@ -321,7 +321,7 @@ defmodule Combo.Template do
   defp default_encoders do
     [
       html: Combo.Template.HTMLEncoder,
-      json: Combo.json_module(),
+      json: Combo.json_library(),
       js: Combo.Template.HTMLEncoder
     ]
   end