+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### `Combo`
+
+- rename `json_module/0` to `json_library/0`, because other packages are using the same pattern, like `:postgrex`, `:swoosh`, etc.
+
 ### `Combo.Proxy`
 
 - sort backends by specificity