+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### `Combo`
 
 - rename `json_module/0` to `json_library/0`, because other packages are using the same pattern, like `:postgrex`, `:swoosh`, etc.
+- change default json_library to `Jason`
 
 ### `Combo.Proxy`